% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{calculate_index_extent}
\alias{calculate_index_extent}
\title{This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent.}
\usage{
calculate_index_extent(clipExtent, h5Extent)
}
\arguments{
\item{clipExtent}{xxxx}

\item{h5Extent}{XXX}
}
\description{
This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent.
}
\examples{
calculate_index_extent(clipExten, h5Extent)

}
\keyword{extent}
\keyword{hdf5,}

