% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{clean_refl_data}
\alias{clean_refl_data}
\title{Clean reflectance data}
\usage{
clean_refl_data(fileName, reflMatrix, epsg)
}
\arguments{
\item{reflMatrix}{, the matrix read in to be converted to a raster.}

\item{epsg}{- the epsg code for the CRS used to spatially locate the raster.}

\item{filename}{the path to the h5 file.}
}
\description{
This function reads in data from the "Reflecatnce" dataset, applies the data
ignore value, scales the data and returns a properly "projected" raster object.
}
\examples{
clean_refl_data(fileName, reflMatrix, epsg)
}
\keyword{extent}
\keyword{hdf5,}

