% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{clean_spectra}
\alias{clean_spectra}
\title{This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent. It returns a rasterStack object of bands.}
\usage{
clean_spectra(spectra, wavelengths)
}
\arguments{
\item{spectra}{REQUIRED. a LIST of spectra returned from Lapply.}

\item{wavelengths}{REQUIRED. The vector of wavelength values of the same length as 
the spectra object}
}
\description{
This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent. It returns a rasterStack object of bands.
}
\examples{
clean_spectra(fileName, bandNum)
}
\keyword{extent}
\keyword{hdf5,}

