% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{create_extent}
\alias{create_extent}
\title{Create h5 file extent ####}
\usage{
create_extent(fileName)
}
\arguments{
\item{filename}{the path to the h5 file}

\item{res}{a vector of 2 objects - x resolution, y resolution}
}
\description{
This function uses a map tie point for an h5 file and data resolution to
create and return an object of class extent.
}
\examples{
create_extent(fileName, res=c(xres, yres))
}
\keyword{extent}
\keyword{hdf5,}

