% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{create_extent_subset}
\alias{create_extent_subset}
\title{Get subset index x,y min/max dimensions}
\usage{
create_extent_subset(h5.extent, dims, res = c(1, 1))
}
\arguments{
\item{h5.extent}{the spatial extent (of class extent) of the H5 file to be subsetted}

\item{dims}{- the x/y min and max INDEX values used to slice the H5 file}

\item{res}{- a vector of 2 objects representing the X and Y resolution (pixel size)}
}
\description{
This function takes an input h5 file extent (the file that you wish to subset
and it creates a subset spatial extent object for the new
subset file (x/y min and max in map units).
}
\examples{
create_extent_subset(h5.extent, dims, res=c(xres, yres))
}
\keyword{dimensions}
\keyword{hdf5,}

