% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{extract_av_refl}
\alias{extract_av_refl}
\title{This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent. It returns a rasterStack object of bands.}
\usage{
extract_av_refl(aRaster, aMask = NULL, aFun = mean)
}
\arguments{
\item{aRaster}{REQUIRED. a raster within an H5 file that you wish to get an mean, max, min value from}

\item{aMask}{REQUIRED. a raster of the same extent with NA values for areas that you don't 
want to calculate the mean, min or max values from.}

\item{aFun}{REQUIRED. the function that you wish to use on the raster, mean, max, min etc}
}
\description{
This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent. It returns a rasterStack object of bands.
}
\examples{
extract_av_refl(aRaster, aMask, aFun=mean)

}
\keyword{extent}
\keyword{hdf5,}

