% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{get_data_dims}
\alias{get_data_dims}
\title{Get Data Dimensions}
\usage{
get_data_dims(fileName)
}
\arguments{
\item{fileName}{a path to the H5 file that you'd like to open}
}
\description{
This function grabs the x,y and z dimemsions of an H5 dataset called "Reflectance"
It would be more robust IF you could pass it the dataset name / path too
}
\examples{
get_data_dims("filename.h5")
}
\keyword{dimensions}
\keyword{hdf5,}

