% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{get_spectra}
\alias{get_spectra}
\title{This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent. It returns a rasterStack object of bands.}
\usage{
get_spectra(fileName, bandNum = 1, epsg = 32611, subset = TRUE,
  dims = NULL, mask = NULL, fun = mean)
}
\arguments{
\item{fileName}{REQUIRED. The path to the h5 file that you want to open.}

\item{bandNum}{REQUIRED. The band number that you wish to open, default = 1}

\item{epsg}{the epsg code for the CRS that the data are in.}

\item{dims, }{an optional object used if subsetData = TRUE that specifies the 
index extent to slice from the h5 file}

\item{mask}{a raster containg NA values for areas that should not be included in the 
output summary statistic.}

\item{fun}{the summary statistic that you wish to run on the data  (e.g. mean, max, min)
default = mean}

\item{subsetData, }{a boolean object. default is FALSE. If set to true, then
... subset a slice out from the h5 file. otherwise take the entire xy extent.}
}
\description{
This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent. It returns a rasterStack object of bands.
}
\examples{
get_spectra(fileName, bandNum)
}
\keyword{extent}
\keyword{hdf5,}

