% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{get_wavelength}
\alias{get_wavelength}
\title{This internal function assigns meaningful names to raster layers in stacks.}
\usage{
get_wavelength(file, bands)
}
\arguments{
\item{file}{the path to the h5 file}

\item{bands}{the band numbers in the reflectance data}
}
\description{
This internal function assigns meaningful names to raster layers in stacks.
}

