% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{open_band}
\alias{open_band}
\title{This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent.}
\usage{
open_band(fileName, bandNum, epsg, subsetData = FALSE, dims = NULL)
}
\arguments{
\item{fileName}{the path to the h5 file that you wish to open.}

\item{bandNum}{the band number in the reflectance data that you wish to open}

\item{epsg}{the epsg code for the CRS that the data are in.}

\item{subsetData, }{a boolean object. default is FALSE. If set to true, then
... subset a slice out from the h5 file. otherwise take the entire xy extent.}

\item{dims, }{an optional object used if subsetData = TRUE that specifies the
index extent to slice from the h5 file}
}
\description{
This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent.
}
\examples{
open_band(fileName, bandNum, epsg, subsetData=FALSE, dims=NULL)

}
\keyword{extent}
\keyword{hdf5,}

