% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-data.R
\name{calculate_index_extent}
\alias{calculate_index_extent}
\title{This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent.}
\usage{
calculate_index_extent(clipExtent, h5Extent, xscale = 1, yscale = 1)
}
\arguments{
\item{clipExtent}{the spatial extent of the clip region that you wan to use to subset data.}

\item{h5Extent}{the spatial extent of the H5 file. If the extents don't overlap this function won't work.}
}
\description{
This function calculates an index based subset to slice out data from an H5 file
using an input spatial extent.
}
\examples{
calculate_index_extent(clipExten, h5Extent)

}
\keyword{extent}
\keyword{hdf5,}

